Extract a base class for the execution pools
Both pools share the capacity ledger: the reserve/restore accounting,
the idle notification, and the reserve-then-schedule shape of post.
Move all of that to ExecutionPools::Base, leaving each pool with just
its scheduling mechanism: the thread pool schedules executions on a
Concurrent::ThreadPoolExecutor via schedule, and the fiber pool hands
them to its reactor, adding its fatal-error and shutdown guards on
top of the base's post.

The default way of performing an execution, wrapped in the app
executor, reporting errors and restoring capacity, also lives in the
base class; the fiber pool overrides it to treat Async::Stop as fatal.

The pool type is derived from the class name, and the factory performs
the lookup in reverse, trusting the type it receives: the worker only
ever passes :thread or :fiber.
